feat: Moves /session path to a version endpoint
