+++ Gemfile
@@ -49,4 +49,6 @@ group :development, :test do
 
   # Use RSpec for testing [https://rspec.info/]
   gem "rspec-rails", "~> 8.0.0"
+
+  gem "guard-rspec", require: false
 end
