+++ config/routes.rb
@@ -1,3 +1,7 @@
 Rails.application.routes.draw do
-  resource :session, only: %i[create]
+  namespace :api do
+    namespace :v1 do
+      resource :session, only: %i[create]
+    end
+  end
 end
